move register extensions to connection
